refacto query >> agent class
